setting up the entity
added constructor for data population in entity
IncomeStatement DTO adjustments for Lombok
BasicData value data type changed from double to Float

+++ src/main/java/com/sawdks/json/stockfinancials/dto/BasicData.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class BasicData {
     private String label;
-    private double value;
+    private Float value;
     private String unit;
     private int order;
 }
